// template/react-starter-kit/src/.generated/api/model/termsControllerGetAdminTermsDirectionItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export type TermsControllerGetAdminTermsDirectionItem = typeof TermsControllerGetAdminTermsDirectionItem[keyof typeof TermsControllerGetAdminTermsDirectionItem];


export const TermsControllerGetAdminTermsDirectionItem = {
  asc: 'asc',
  desc: 'desc',
} as const;