// template/react-starter-kit/src/.generated/api/model/termsControllerGetAdminTermsSortItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export type TermsControllerGetAdminTermsSortItem = typeof TermsControllerGetAdminTermsSortItem[keyof typeof TermsControllerGetAdminTermsSortItem];


export const TermsControllerGetAdminTermsSortItem = {
  publishedAt: 'publishedAt',
  createdAt: 'createdAt',
  version: 'version',
  id: 'id',
} as const;