// template/react-starter-kit/src/.generated/api/model/updateTermGroupRequestDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateTermGroupRequestDto {
  /** @maxLength 50 */
  code?: string;
  /** @maxLength 255 */
  title?: string;
  isRequired?: boolean;
  sortOrder?: number;
}